+++ src/controllers/SendoutsController.php
@@ -6,7 +6,6 @@
 namespace putyourlightson\campaign\controllers;
 
 use Craft;
-use craft\elements\User;
 use craft\errors\ElementNotFoundException;
 use craft\helpers\DateTimeHelper;
 use craft\web\Controller;
@@ -249,25 +248,14 @@ class SendoutsController extends Controller
             'status' => ContactElement::STATUS_ACTIVE,
         ];
 
-        // Get test contact
-        $variables['testContact'] = null;
+        // Get test contacts
+        $variables['testContacts'] = [];
 
-        if ($sendout->campaignId !== null) {
-            $campaign = $sendout->getCampaign();
+        $campaign = $sendout->getCampaign();
 
-            if ($campaign !== null) {
-                $campaignType = $campaign->getCampaignType();
-
-                if ($campaignType->testContactId !== null) {
-                    $variables['testContact'] = Campaign::$plugin->contacts->getContactById($campaignType->testContactId);
-                }
-            }
-        }
-
-        if ($variables['testContact'] === null) {
-            /** @var User|null $currentUser */
-            $currentUser = Craft::$app->user->getIdentity();
-            $variables['testContact'] = $currentUser ? Campaign::$plugin->contacts->getContactByEmail($currentUser->email) : null;
+        if ($campaign !== null) {
+            $campaignType = $campaign->getCampaignType();
+            $variables['testContacts'] = $campaignType->getTestContactsWithDefault();
         }
 
         // Determine which actions should be available
@@ -513,22 +501,20 @@ class SendoutsController extends Controller
         $this->requirePostRequest();
         $this->requireAcceptsJson();
 
-        $contactId = Craft::$app->getRequest()->getBodyParam('contactId');
+        $contactIds = Craft::$app->getRequest()->getBodyParam('contactIds');
         $sendout = $this->_getSendoutFromParamId();
 
-        // Validate test contact
-        if (empty($contactId)) {
-            return $this->asJson(['success' => false, 'error' => Craft::t('campaign', 'A contact must be submitted.')]);
+        // Validate test contacts
+        if (empty($contactIds)) {
+            return $this->asJson(['success' => false, 'error' => Craft::t('campaign', 'At least one contact must be submitted.')]);
         }
 
-        $contact = Campaign::$plugin->contacts->getContactById($contactId);
+        $contacts = Campaign::$plugin->contacts->getContactsByIds($contactIds);
 
-        if ($contact === null) {
-            return $this->asJson(['success' => false, 'error' => Craft::t('campaign', 'Contact not found.')]);
-        }
-
-        if (!Campaign::$plugin->sendouts->sendTest($sendout, $contact)) {
-            return $this->asJson(['success' => false, 'error' => Craft::t('campaign', 'Couldn’t send test email.')]);
+        foreach ($contacts as $contact) {
+            if (!Campaign::$plugin->sendouts->sendTest($sendout, $contact)) {
+                return $this->asJson(['success' => false, 'error' => Craft::t('campaign', 'Couldn’t send test email.')]);
+            }
         }
 
         return $this->asJson(['success' => true]);