+++ src/resources/js/CampaignEdit.js
@@ -17,8 +17,13 @@ Campaign.CampaignEdit = Garnish.Base.extend(
 
             $('.send-test').addClass('disabled');
 
+            var contactIds = [];
+            $('#testContacts input').each(function() {
+                contactIds.push($(this).val());
+            });
+
             var data = {
-                contactId: $('#testContact input').val(),
+                contactIds: contactIds,
                 campaignId: $('input[name=campaignId]').val()
             };